[cfid-397] Add approvals to ScimUser.
In addition, relax authorizations for /Users such that users can access their own ScimUser
resource. Also fix bug cfid-487. Corresponding pivotal tasks: [#41051829] and [#42174475].

+++ common/src/test/java/org/cloudfoundry/identity/uaa/oauth/authz/ApprovalsAdminEndpointsTests.java
@@ -86,7 +86,7 @@ public class ApprovalsAdminEndpointsTests {
 		endpoints.setSecurityContextAccessor(mockSecurityContextAccessor(marissa.getId()));
 	}
 
-	private void addApproval(String userName, String clientId, String scope, long expiresIn, ApprovalStatus status) {
+	private void addApproval(String userName, String clientId, String scope, int expiresIn, ApprovalStatus status) {
 		dao.addApproval(new Approval(userName, clientId, scope, expiresIn, status));
 	}
